win: Fix for logging.h ERROR, again
I didn't think the previous attempt through (maybe this is what you
were referring to in the review). The problem is #undef ERROR is
insufficient because logging.h avoids including windows.h. The #undef
would work if windows.h were included by logging.h or when it happens
to be included before logging.h in a given translation unit, but not
if it's included afterwards.

Instead, define ERROR up front to match the value of windows.h and
add LOG_0, et al. to handle the undesirable macro expansions.

R=mark@chromium.org
BUG=crashpad:1

Review URL: https://codereview.chromium.org/801643010