win: Default build to x64
For Windows Crashpad's purposes at the moment, an x64 build is more
useful so we can test cross-bitness access. We'll probably need to
support an x86 build at some point for x86 host OS, but that can wait
until some time in the future as we'll need to get x86 OS bots and
fiddle with the waterfall configurations.

R=mark@chromium.org
BUG=crashpad:1

Review URL: https://codereview.chromium.org/981333002