Clean up -Wunused-parameter warnings reachable by Crashpad clients.
We aren’t concerned with -Wunused-parameter in our own code or in
Crashpad, and disable this warning in build/common.gypi. However, some
mini_chromium headers are exposed to consumers of the Crashpad client
library, and thus may build with -Wunused-parameter in effect. In clang,
this warning is included in -Wextra.

To make it easier for consumers to use Crashpad, clean up the two minor
things that violate -Wunused-parameter in exposed headers.

As discussed in
https://groups.google.com/a/chromium.org/d/msg/crashpad-dev/NaB7PrfW04g/s-cOkTiTG_wJ.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/1026703003