Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I hope it doesn’t need to be custom. Perhaps we can get by with a little more stubbing in mini_chromium.

Yeah, maybe that could work too. I was thinking it needed to be custom because it actually needs code from //components/metrics as well as //base. But maybe we can wrap that up so there's a more focused thing in base/mini_chromium that Crashpad can call.

FWIW, I don't think it's too bad with a custom main, as I *think* it's really only "main() { crashpad::HandlerMain(); }" that gets replaced.

Patch-set: 4