Update patch set 2
Patch Set 2:

Using the packaged vs_toolchain is a horrible mess of pulling in most of build/ from Chrome, but we can do a simpler version that should be enough for Crashpad.

TBD is how to have something autodetect the VS installation and run and extract the environment automatically (right now it'd have to be run manually before running GN). The GN stuff chromium/src/build/ currently depends on gyp to do that, a dependency that I don't have much desire to duplicate ಠ_ಠ.

Patch-set: 2