android: Don’t set _FILE_OFFSET_BITS, it never worked
off_t is normally 32 bits on 32-bit platforms. It’s customary to set
_FILE_OFFSET_BITS to 64 to get a 64-bit off_t on these platforms, but
this was traditionally never effective in the Android NDK, so the macro
was being set without effect. That’s changed with “unified headers,” but
choosing a 64-bit off_t results in certain functions without historical
64-bit off_t support becoming unavailable below a certain API level. For
example, mmap() and sendfile() aren’t available with a 64-bit off_t
until API 21.

64-bit platforms always get a 64-bit off_t, and this macro is
unnecessary.

Bug: crashpad:30
Test: build for a 32-bit platform using API 16 and unified headers
Change-Id: Ib4e422c480c5360afbb5638427150a029bdc052c
Reviewed-on: https://chromium-review.googlesource.com/478371
Reviewed-by: Joshua Peraza <jperaza@chromium.org>