Update patch set 7
Patch Set 7:

It may be better to not try to cross this road in this change, but how is sysroot intended to interact with cross-compiles? In Chromium, it looks like the tunable argument may just be target_sysroot, so I guess some additional attention is needed.

Patch-set: 7
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>