gn win: Automatically find location of Visual Studio
Previously, a vcvarsall or equivalent command was required to have been
run before GN, and the build files did not attempt to locate the Visual
Studio installation. Additionally, the tools in the PATH were assumed to
be compatible with the setting of target_cpu as given to GN, and the
build would confusingly error out otherwise.

Now, at GN-time, the location of the VS2015 install is found, the
required environment extracted from vcvarsall and saved, and this
environment is used when invoking the compiler, linker, and assembler.
Unfortunately this requires a bunch of goop in win_helper.py (mostly
from old Chromium).

Currently, the x64 toolchain is always used. In a follow up change, the
toolchain definition should be templated and the correct one chosen
based on the value of target_cpu.

Bug: crashpad:79
Change-Id: Id6150f3a80b9f26154b30312605989b89e479313
Reviewed-on: https://chromium-review.googlesource.com/905572
Reviewed-by: Mark Mentovai <mark@chromium.org>