mac-arm64: Allow target_cpu = "mac_universal" to create universal builds
When building for macOS and configured with target_cpu =
"mac_universal", bi-architecture x86_64/arm64 output will be produced.

mac_universal is, so far, mini_chromium-specific, and does not mirror
any supported Chromium configuration. It exists to support Keystone,
which intends to ship as x86_64/arm64 universal, and requires a
universal build of Crashpad.

Bug: crashpad:345
Change-Id: If14ced0bfb8ca83899e180ccabcba5211fe674fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/mini_chromium/+/2386603
Reviewed-by: Robert Sesek <rsesek@chromium.org>