Add base::StrCat().
This also requires adding a version of base::span, lest I want to use
std::span from C++20 directly instead (I do not). I cut down the usage
of additional templates/utilities from the Chromium version somewhat to
try to keep this span implementation more standalone.

In adding the necessary bits to template_util.h, I discovered no one is
using the existing file, so I gutted it. (Upstream, something similar
was done in 2016, when ~the whole file was obsoleted in favor of
<type_traits> before slowly growing anew.)

Bug: 1371963
Change-Id: Ie680b254feaa68e3f921fc3f397dd38780fa5a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/mini_chromium/+/4951093
Reviewed-by: Mark Mentovai <mark@chromium.org>