Fix win-arm64 MSVC builds, and maybe other builds
A GN conditional was accidentally structured:

```
if (invoker.current_cpu == "arm64") {
  ml = "armasm64.exe"
} if (invoker.current_cpu == "x64") {
  ml = "ml64.exe"
} else {
  ml = "ml.exe"
}
```

The second `if` should have been `else if`.

The file was not `gn format`ted or `git cl format`ted. Formatting would
have made the error more readily apparent during code review and
subsequent troubleshooting, as there would not have been a tendency to
misread the second `if` as `else if`, making it much more evident that
in the arm64 case, `ml` would have been reset to `ml.exe`:

```
if (invoker.current_cpu == "arm64") {
  ml = "armasm64.exe"
}
if (invoker.current_cpu == "x64") {
  ml = "ml64.exe"
} else {
  ml = "ml.exe"
}
```

Bug: 399577698
Change-Id: I390c96fe8436ecaa540fe9460d6e8582ad29e7db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/mini_chromium/+/6308964
Reviewed-by: Nico Weber <thakis@chromium.org>