build: Add tvOS's id to BuildInfo.plist's UIDeviceFamily array
Not doing so causes the generated bundles (e.g. Crashpad tests) to
fail to run:

```
Unable to Install “crashpad_client_test”
Domain: IXUserPresentableErrorDomain
Code: 6
Failure Reason: This app is not made for this device.
Recovery Suggestion: This app was not built to support this device family; app is compatible with (
    1,
    2
) but this device supports (
    3
)
```

Bug: chromium:405140658
Change-Id: I9612263c3c90bf22b8084673c3f5f34dbd107f87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/mini_chromium/+/6383575
Reviewed-by: Justin Cohen <justincohen@chromium.org>