build: Add tvOS support to C++ and GN
This is the mini_chromium version of https://crrev.com/c/6306066,
which added tvOS support to //build in Chromium.

Like in Chromium, tvOS is treated like a subset of iOS: a build is
considered to be a tvOS build when a regular iOS build's
`target_platform` GN variable is set to "tvos" (by default its value
is "iphoneos"). In other words, `target_os` is still "ios", and
`is_ios` is true.

On the C++ side, BUILDFLAG(IS_IOS_TVOS) is also set if and only if
BUILDFLAG(IS_IOS) is also true (which mimics Apple's own
TargetConditionals.h hierarchy given the variables we use to set the
other IS_* flags).

Bug: chromium:405140658
Change-Id: I3acd1f4c9226789b708e586d9f50052398678884
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/mini_chromium/+/6382038
Reviewed-by: Justin Cohen <justincohen@chromium.org>